Replace FAB with BoardActionButtons component
Create reusable BoardActionButtons component with Add dropdown (New
Phrase/New Board), Reader, and Edit/Done buttons. Use it in both
SwipeableBoardNavigator (mobile) and BoardSelector (desktop).

Remove PhrasesActionMenu FAB component entirely since the actions are
now accessible via the header buttons on both mobile and desktop.

Closes #227
